requirements.txt
# Core dependencies for Claims Health Analytics Agents
python-dotenv>=1.0.0
azure-identity>=1.15.0
semantic-kernel>=1.37.0
aiohttp>=3.9.0

# Azure AI Services dependencies
azure-core>=1.29.0
azure-ai-textanalytics>=5.3.0
azure-ai-formrecognizer>=3.3.0
azure-ai-documentintelligence>=1.0.0
azure-ai-inference>=1.0.0b9
openai>=1.35.0

# OpenTelemetry tracing dependencies
opentelemetry-instrumentation-openai-v2==2.1b0
opentelemetry-sdk==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1

# Additional dependencies for document processing and health analytics
Pillow>=10.0.0
pypdf>=4.0.0
python-docx>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Optional dependencies for enhanced functionality
typing-extensions
dataclasses-json